+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Use Sass (libsass) instead of Compass
 * Unminify CodeMirror scripts
 * Fixes for the WP 3.8 interface
+* Fix 'enable snippets menu for site admins' multisite setting
 
 ## 1.9
 * Add and remove network capabilities as super admins are added and removed